use fsspec's own loop
